requirements.txt
annotated-types==0.6.0
anyio==4.0.0
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
copier==8.3.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
dunamai==1.19.0
evdev==1.4.0
exceptiongroup==1.1.3
eyeD3==0.8.10
filelock==3.6.0
flet==0.10.3
flet-core==0.10.3
flet-runtime==0.10.3
funcy==2.0
grpcio==1.30.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
kazam==1.4.5
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
mysql-connector-python==8.1.0
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.2
onboard==1.4.1
packaging==23.2
PAM==0.4.2
pathspec==0.11.2
pexpect==4.8.0
Pillow==9.0.1
plumbum==1.8.2
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pydantic==2.4.2
pydantic_core==2.10.1
pyelftools==0.27
Pygments==2.16.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pypng==0.20220715.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
repath==0.9.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
typing_extensions==4.8.0
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.8
webencodings==0.5.1
websocket-client==1.6.4
websockets==11.0.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
